testUtils - adding getMonthTitle; moving driver to tests folder; moving calendar tests and driver to tests directory
